Python 5/Esame 07.01.25/Rest_App/Esame Python 5 GdG/rest_flask/requirements.txt
Flask
psycopg2-binary
Flask-Session